Release 1.1.0
# Known Issues

- half8 "equals" and "not equals" operators don't conform to the IEEE 754 standard - Unity has not yet reacted to my issue.

# Fixes

- Fixed a bug where vshl-/vshr-ing a byte16 vector by 11 would return the vector itself

# Additions

### RNG
- Added/Modified 8, 16, 32 and 64 bit XOR-Shift pseudo random number generators:
- They use the most efficient (Avx2) SIMD instructions to generate vectors with elements of the corresponding size in bytes. When compared to Unity.Mathematics, the performance is better since a) scalar multiplication of each generated value has been replaced by a single SIMD instruction and b) doubles are generated by Random64 instead of two 32-bit RNG iterations
- Removed NextT(T max) from SIGNED integer and floating point types, since those will never generate negative numbers. One can either generate an unsigned integer and cast it to a signed value for free, or use the functions with min and max parameters, as both of these would be more clear in regards to what range the result will be in
- Unity.Mathematics.Random is implicitly convertible to Random32 and vice versa. Safe and fast explicit type conversions between Random8, 16, 32 and 64 were added

### Shuffle

- Added bool8, bool16 and bool32 subvector getters
- Added half8 setters
- Added (s)byte32 getters
- Added setters for (s)byte8 and s)byte16 subvectors
- Added setters for (u)short8 subvectors
- !!! Setters for (s)byte32, (u)short16, (u)int8 and float8 subvectors are implemented, but due to Unity.Burst related bugs, they are deactivated in the code. The issue has been forwarded a month ago and should be fixed with Burst 1.5

- Slightly improved the performance of a select few byte2/3/4 vector shuffle getters and setters

+++ Runtime/Functions/Comparison/Count.cs
@@ -11,40 +11,40 @@ namespace MaxMath
     {
         /// <summary>       Returns the number of true values in a bool2 vector.        </summary>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]  [return: AssumeRange(0, 2)]
-        public static int count(bool2 x)
+        public static uint count(bool2 x)
         {
 Assert.IsSafeBoolean(x.x);
 Assert.IsSafeBoolean(x.y);
 
-            return math.countbits((int)(*(ushort*)&x));
+            return (uint)math.countbits((uint)(*(ushort*)&x));
         }
 
         /// <summary>       Returns the number of true values in a bool3 vector.        </summary>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]  [return: AssumeRange(0, 3)]
-        public static int count(bool3 x)
+        public static uint count(bool3 x)
         {
 Assert.IsSafeBoolean(x.x);
 Assert.IsSafeBoolean(x.y);
 Assert.IsSafeBoolean(x.z);
 
-            return math.countbits(0x00FF_FFFF & *(int*)&x);
+            return (uint)math.countbits(0x00FF_FFFF & *(int*)&x);
         }
 
         /// <summary>       Returns the number of true values in a bool4 vector.        </summary>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]  [return: AssumeRange(0, 4)]
-        public static int count(bool4 x)
+        public static uint count(bool4 x)
         {
 Assert.IsSafeBoolean(x.x);
 Assert.IsSafeBoolean(x.y);
 Assert.IsSafeBoolean(x.z);
 Assert.IsSafeBoolean(x.w);
 
-            return math.countbits(*(int*)&x);
+            return (uint)math.countbits(*(int*)&x);
         }
 
         /// <summary>       Returns the number of true values in a bool8 vector.        </summary>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]  [return: AssumeRange(0, 8)]
-        public static int count(bool8 x)
+        public static uint count(bool8 x)
         {
 Assert.IsSafeBoolean(x.x0);
 Assert.IsSafeBoolean(x.x1);
@@ -55,12 +55,12 @@ Assert.IsSafeBoolean(x.x5);
 Assert.IsSafeBoolean(x.x6);
 Assert.IsSafeBoolean(x.x7);
 
-            return math.countbits(*(long*)&x);
+            return (uint)math.countbits(*(long*)&x);
         }
 
         /// <summary>       Returns the number of true values in a bool16 vector.        </summary>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]  [return: AssumeRange(0, 16)]
-        public static int count(bool16 x)
+        public static uint count(bool16 x)
         {
 Assert.IsSafeBoolean(x.x0);
 Assert.IsSafeBoolean(x.x1);
@@ -79,12 +79,12 @@ Assert.IsSafeBoolean(x.x13);
 Assert.IsSafeBoolean(x.x14);
 Assert.IsSafeBoolean(x.x15);
 
-            return math.countbits(Sse2.movemask_epi8(Sse2.slli_epi16(x, 7)));
+            return (uint)math.countbits(Sse2.movemask_epi8(Sse2.slli_epi16(x, 7)));
         }
 
         /// <summary>       Returns the number of true values in a bool32 vector.        </summary>
         [MethodImpl(MethodImplOptions.AggressiveInlining)]  [return: AssumeRange(0, 32)]
-        public static int count(bool32 x)
+        public static uint count(bool32 x)
         {
 Assert.IsSafeBoolean(x.x0);
 Assert.IsSafeBoolean(x.x1);
@@ -119,7 +119,7 @@ Assert.IsSafeBoolean(x.x29);
 Assert.IsSafeBoolean(x.x30);
 Assert.IsSafeBoolean(x.x31);
 
-            return math.countbits(Avx2.mm256_movemask_epi8(Avx2.mm256_slli_epi16(x, 7)));
+            return (uint)math.countbits(Avx2.mm256_movemask_epi8(Avx2.mm256_slli_epi16(x, 7)));
         }
     }
 }
